+++ app/src/main/java/kosenda/makecolor/navigation/Navigation.kt
@@ -36,12 +36,12 @@ import kosenda.makecolor.feature.display.screen.SplitColorScreen
 import kosenda.makecolor.feature.display.screen.SplitScreen
 import kosenda.makecolor.feature.edit.RegisterScreen
 import kosenda.makecolor.feature.info.InfoDialog
-import kosenda.makecolor.feature.makecolor.screen.InputTextScreen
-import kosenda.makecolor.feature.makecolor.screen.MergeScreen
-import kosenda.makecolor.feature.makecolor.screen.PickerScreen
-import kosenda.makecolor.feature.makecolor.screen.PictureScreen
-import kosenda.makecolor.feature.makecolor.screen.RandomScreen
-import kosenda.makecolor.feature.makecolor.screen.SeekbarScreen
+import kosenda.makecolor.feature.screen.InputTextScreen
+import kosenda.makecolor.feature.screen.MergeScreen
+import kosenda.makecolor.feature.screen.PickerScreen
+import kosenda.makecolor.feature.screen.PictureScreen
+import kosenda.makecolor.feature.screen.RandomScreen
+import kosenda.makecolor.feature.screen.SeekbarScreen
 import kosenda.makecolor.feature.settings.SettingScreen
 import kosenda.makecolor.ui.GoogleAd
 import kotlinx.serialization.decodeFromString
