+++ feature/makecolor/src/main/java/kosenda/makecolor/feature/viewmodel/InputTextViewModel.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.feature.makecolor.viewmodel
+package kosenda.makecolor.feature.viewmodel
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
