+++ feature/makecolor/src/main/java/kosenda/makecolor/feature/viewmodel/PictureViewModel.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.feature.makecolor.viewmodel
+package kosenda.makecolor.feature.viewmodel
 
 import android.content.Context
 import android.graphics.Bitmap
